Record VRatelimit metrics
With these metrics, we'll be able to see the number of times:

1. We set a user-based limit
2. We set an ip-based limit
3. A user is exempted from ratelimit checks
4. We check a user-based limit
5. We check an ip-based limit
6. A user-based limit is hit
7. An ip-based limit is hit

Dealing with sampling rates is a little bit tricky, since we're recording
per-prefix and some endpoints are hit much more frequently than others.  I did
a simplistic grep to get the endpoints that use `VRatelimit`, then took a look
at the rate of requests for all of those; it looks like a 10% sampling rate
shouldn't overwhelm our graphite servers on the popular endpoints, and won't be
*awful* for the infrequently-requested ones (as long as we look at long time
periods for them).