Revert "Comment sort preference: dual-write to migrate"
This reverts commit c3e9879cac6d50df08971629c07bc4c84e93eb1d.

It's now been over a month since we launched the new comment sort preference.
We've long ago hit the long tail of accounts needing a migration from the old
preference to the new one, and the ones that remain are either very inactive or
completely inactive (lost password, etc.), and thus, shouldn't really be
bothered much by getting put back on the default sort of "best" (`confidence`).