Q&A sort: always show OP replies
As a few users have pointed out, it doesn't make much sense for OP replies to
be hidden in Q&A sort due to downvotes, since the entire point of the sort is
to make it easier to find what they're saying.

So now we'll override the comment score threshold for those particular comments
while in Q&A sort.