AutoMod: fix bug with "skipped" search checks
The variable that holds the result of attempting to match the search
check was not being reset for each individual check, so if a check was
skipped due to its fields being discarded, it would keep the same value
it had from a previous check.

In practice, this means that a check like this will not match on a link
submission to baddomain.com:

    url+body: "baddomain.com"
    ~body: "I'm an exception"

The match from the url check would still be present after the ~body
check was skipped, so this would end up causing the rule to be
considered unsatisfied, since the ~body check would seem to have matched
successfully.