Orangereds as emails: Add cooldown period for batching
If a notification is created within the cooldown period, then wait a little bit longer
to grab messages within a similar timeframe before sending the email. The max_delay ensures
that it won't remain in the batching state for too long, but the user might receive multiple
batches in the event of frequent notifications.