Add support for multiple stickies (set to 2)
This includes adding a function to Subreddit for get_sticky_fullnames()
in order to help the transition from the previous sticky_fullname
attribute. In the future this can most likely be removed and we can just
use the new sticky_fullnames attribute directly.

The API endpoint also adds support for stickying into a specific slot
(which will replace any existing sticky in that slot, if any). This is
not currently exposed in the web interface.