Subreddit: disable memoized contributor_ids().
UserRel will by default keep a memoized list of all accounts in a
given relation (e.g. SRMember "contributor"). This list is never used
anywhere, and is very expensive to keep updated. Every time a relation
is created or destroyed a query is executed to get ALL accounts in the
relation so the list can be updated.