Subreddit._by_name: Drop an unnecessary join on the Thing table.
Since we only need to get the subreddit ID from its name, we don't care
about anything in the Thing table.  The implicit deleted=True rule and
lack of optimize_rules meant that extra rules were being added to the
query which made a join on the Thing table necessary.  This should make
a cold cache scenario a little less painful.