Subreddit._by_name: Cache negative lookups.
A huge portion of the queries to the subreddit databases are _by_name
lookups for subreddits that don't exist.  By caching negative lookups,
we can hopefully reduce some portion of these.  This is somewhat of an
experiment though because

1) it assumes that there will be a decent hit
rate on the negative lookups, i.e. that they're not a super long-tail
distribution of crazy names, and

2) that the cached negative lookups won't have worse effects on the
system by filling the caches with negative results which leave less room
for "real" data.