stats: Don't swallow exceptions in timer context manager.
The return value of the __exit__ function of a context manager is only
used for its truthiness to determine if any exceptions within the "with"
block should be swallowed.

From PEP-0343:

    IMPORTANT: if mgr.__exit__() returns a "true" value, the exception
    is "swallowed".

Since the timer was returning "self" which evaluates truthily, any
exceptions inside a timer "with" block would be swallowed. Scary!