Make `_wsf()` keep the original spacing by default
The space compressor has a habit of changing the semantics of text,
and we generally work around it by `text.replace(" ", "&#32;")`.

We know that the format string passed to `_wsf()` won't contain HTML,
so we don't have to worry about mucking with spaces between attributes.
Just automatically work around the space compressor by default.