Update messaging labels to try to reduce confusion
There has been a long-standing issue with users sending modmail when
they're trying to make a new text post to the subreddit. From examining
a number of examples of this happening, it appears that they tend to do
this by going to a message page (inbox/unread) by clicking the envelope
icon, then they click the "compose" tab, and type the name of the
subreddit into the "to" box.

This commit updates several labels to make it more clear that it's a
private message that's going to be sent, not "posting a message
publicly".