TEMP: Appease pickle by creating a stub `r2.lib.cookies`
Pagecache is shared between apps, and it contains pickled cookies.
If a new app stores a pickle with a reference to
`r2.lib.cookies.Cookie` in the middle of the push, the apps that
haven't been updated yet will lose their minds.

Work around that by importing `Cookie` and `Cookies` into a stub
`r2.lib.cookies` and pushing it before the main changeset.