setup.py: Remove check for distribute>=0.6.16.
This check was in place for d6f91eb1b4d9026b8012d3645ba4257151b82a8a to make
sure that we were using a new enough version of distribute to handle
the "cythonize" command. distribute was a now-deprecated fork of setuptools,
and the versions of setuptools on precise and trusty are new enough to not
need this check.