Mark NSFW when submitting if title matches regex
This is the first step in trying to fix a long-standing annoyance with
the NSFW toggling: inability to toggle the NSFW flag off if the post has
"NSFW" or "NSFL" in its title. This is because the title check is
currently done at display-time, instead of just doing the check at
submit-time and have the flag default to enabled if the title matches.

In order to fix this, we'll need to do it in three steps:

1. Start marking NSFW at submit-time if the title matches (this commit)
2. Do a backfill to mark all past posts as NSFW if their title matches
3. Turn off the display-time check