DO NOT OPENSOURCE: temporary check for NSFW titles
We want to validate that my backfill of setting the NSFW flag for
everything with "NSFW" or "NSFL" in the title was done correctly, so
this will send me a private message on Slack if anything has a mismatch
between the display-time title check and whether the flag is set or not.
We can leave this on for a few days and then I'll be able to remove the
display-time check if there aren't any errors.

Note that this only checks for things where the title check comes back
positive but the flag isn't set, not vice-versa. We can't check in that
direction since it's already possible for people to manually add the
flag for things where the title doesn't match. Checking for this sort of
mismatch is the more important one anyway, since it would mean that
removing the display-time check would end up un-marking some things as
NSFW that previously were, which is definitely the more dangerous
possibility.