Feature flags: add support for experiments
We've long wanted the ability to run A/B tests; they make it much easier to
test out potential behavior, and reduce the self-selection bias you get through
opt-in betas.  This commit adds A/B-type experiment support to the feature
flagging system.

Currently, only server-side checks and tests on logged-out users are directly
supported.  This is purely to scope down the feature enough to make it easier
to get a v1 out.