Timeouts: hide report reasons for mods in timeout
We are currently stopping (via modal) mods in timeout from viewing report
reasons.  Now we will also be hiding that data in the api as well.

Additionally, to be consistent, we're removing the data from the html of
standard pages, even if the modal blocks easy access to it.

The number of reports is left alone, also to maintain consistency with the
desktop UI.