Remove account restrictions for app developers
Ever since we added OAuth developers in reddit/reddit@cc40171, we've been
silently filtering out deleted and spammed users in any of the developer
listings.  Recently, we added users in timeout to that list as well.

This can cause some unexpected behavior at times.  One of the most notable is
the check of the developer list when we remove a dev: if there aren't any more
devs, we delete the app.  If a developer of a popular app gets themselves in a
temporary timeout or accidentally gets spammed, this would break the app for
all its users, which is less than ideal.

We discussed this a bit and decided there's not a good reason to tie these
things together.  If a user needs disciplining, that shouldn't affect their
app.  If a user is a spammer, that shouldn't affect their apps (because you
can't do much with that anyways).  And if we actually do want to remove a
user's developer access to an app (say, because they're using it to create
massive amounts of spam), we should do that with a separate tool.

I did leave the filtering out of deleted users in place.  Partly this was
because otherwise a number of other places in the code would need to add
various conditionals to deal with the situation, and I was lazy, but mostly
it's because we shouldn't have any of these, anyways: account deletion removes
the user from all of their apps, and you can't add a deleted user back as a new
developer.