Frontend: Add errors.js
A js module for handling api errors and client-side validation errors.
Right now, most of this is done through the jQuery-json response and
through custom code on a per-feature basis.  This module provides a way
to consume API errors that would normally be displayed through the
jQuery-json response handler, as well as a way to define errors on the
client side that display in the same way.