up/downvoted queries: fix ordering, reduce updates
The voting refactor had an error here - previously, a user's
upvoted/downvoted pages on their profile were sorted based on the time
that the user voted on the post - that is, the most recently-voted
item would be at the top. After the update, the ordering switched to be
based on when the post was made.

This fixes that unintended change, and also reduces the number of query
updates, since "we don't know if it's a changed vote" is no longer true.