Q&A sort: fix comment duplication bug
We were noticing an issue with Q&A sort threads where admin-distinguished
non-OP comments were showing up both in the proper place as well as like
top-level comments.  The root of this was that while those comments were
skipped from being auto-collapsed, this didn't affect any of their ancestors.
This caused the comment builder to remove the ancestors from consideration, so
when it came to the distinguished comment, we thought it was a top-level
comment (because we had deleted its parent from the "in consideration" set).

There are a couple of ways to handle this, but the easiest was to expand
ancestor-expansion to all comments that are immune from auto-collapse.  It also
makes sense from a product perspective.