Deal with bid in dollars in UI
We had previously misnamed the bid amount in sponsored.js. It was supposed
to be in pennies, but it was actually in dollars, which makes sense,
because the fields are expecting dollar amounts. So, instead of making
the pennies-to-dollars conversion in the UI, then making the
dollars-to-pennies conversion again in the controller, let's just deal
with dollars in the UI.