EU Cookies: Migrate to v2 cookie to normalize paths across all cookies.
We accidentally released EU cookies to production setting a cookie
on the current path, which meant that it would show up always for
different paths. This is bad because the cookie has essentially been
made permanently unreliable, such that if we don't know the path that it
was set upon we can't clear it or update it.

Here we're migrating to a v2 of the cookie, but still reading from the
old cookie if it exists so that we don't show it to users if they've
already accepted it from the homepage. If they see it this time, it will
write to the new cookie and on next page load it will take the new
cookie as the preference, which will always be set at the root path.

Luckily the eu_cookie will go away over time,  as it was only a session
cookie at the time it was set, so probably sometime in early 2016 we can
migrate back to just eu_cookie.