VRatelimit: add 'fatal' flag
Many of the other validators have a 'fatal' flag that controls whether errors
will be added to some sort of object that the controller can access, or if
they'll just cause a failed request immediately.  This is useful when the
controller doesn't want to do any special behavior and would just abort for
each error case.