Sticky Comments: Only color titles for stickied links
Previously this was also including stickied comments, which had a
side effect of coloring titles of links on user pages where the
comment was stickied.

This cleans that up, and additionally cleans up the selectors which
were unnecessarily specified for .visited/:visited - the specificity
of this selector overrides .visited/:visited anyhow.